Extract `ByteLineStream` state into submod to refactor.
